% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot Histogram}
\usage{
plot_histogram(df, column, bins = 30)
}
\arguments{
\item{df}{A data frame.}

\item{column}{The name of the column to plot.}

\item{bins}{Number of bins for the histogram.}
}
\value{
A ggplot object.
}
\description{
This function creates a histogram for a specified column in a data frame.
}
\examples{
data(iris)
plot_histogram(iris, 'Sepal.Length')
}
