% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{simple_classification}
\alias{simple_classification}
\title{Simple Classification with Train-Test Split}
\usage{
simple_classification(
  df,
  response,
  predictors,
  learner_name = "classif.rpart",
  train_ratio = 0.7
)
}
\arguments{
\item{df}{A data frame.}

\item{response}{The name of the response variable.}

\item{predictors}{A vector of predictor variable names.}

\item{learner_name}{The name of the learner to use. Default is "classif.rpart".}

\item{train_ratio}{The proportion of data to use for training. Default is 0.7.}
}
\value{
A list with the trained model and performance metrics.
}
\description{
This function fits a classification model using mlr3 with a train-test split.
}
\examples{
data(iris)
simple_classification(iris, 'Species', c('Sepal.Length', 'Sepal.Width', 'Petal.Length', 'Petal.Width'))
}
