% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{simple_lm}
\alias{simple_lm}
\title{Simple Linear Regression with Train-Test Split}
\usage{
simple_lm(df, response, predictor, train_ratio = 0.7)
}
\arguments{
\item{df}{A data frame.}

\item{response}{The name of the response variable.}

\item{predictor}{The name of the predictor variable.}

\item{train_ratio}{The proportion of data to use for training. Default is 0.7.}
}
\value{
A list with the model, training performance, and test performance.
}
\description{
This function fits a simple linear regression model using mlr3 with a train-test split.
}
\examples{
data(mtcars)
simple_lm(mtcars, 'mpg', 'hp')
}
