% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summary_stats}
\alias{summary_stats}
\title{Summary Statistics}
\usage{
summary_stats(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with summary statistics.
}
\description{
This function calculates summary statistics for a given data frame.
}
\examples{
data(iris)
summary_stats(iris)
}
